refactor : 도메인 관련 클래스 패키지 수정
