+++ src/main/java/com/pullanner/api/service/article/ArticleService.java
@@ -6,8 +6,8 @@ import com.pullanner.api.controller.article.dto.ArticleResponse;
 import com.pullanner.api.controller.article.dto.ArticleResponses;
 import com.pullanner.api.controller.article.dto.ArticleSaveRequest;
 import com.pullanner.api.controller.article.dto.ArticleUpdateRequest;
-import com.pullanner.domain.user.entity.User;
-import com.pullanner.domain.user.repository.UserRepository;
+import com.pullanner.domain.user.User;
+import com.pullanner.domain.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
