+++ src/main/java/com/pullanner/global/auth/oauth2/dto/OAuthAttributes.java
@@ -1,7 +1,7 @@
 package com.pullanner.global.auth.oauth2.dto;
 
-import com.pullanner.domain.user.entity.Role;
-import com.pullanner.domain.user.entity.User;
+import com.pullanner.domain.user.Role;
+import com.pullanner.domain.user.User;
 import com.pullanner.global.auth.oauth2.exception.UnsupportedOAuth2ProviderException;
 import java.util.Map;
 import lombok.Builder;
